Trying just applying library support for Unions as is
