+++ codegen/smithy-dafny-codegen/src/main/java/software/amazon/polymorph/smithyrust/generator/RustLibraryShimGenerator.java
@@ -933,80 +933,6 @@ public class RustLibraryShimGenerator extends AbstractRustShimGenerator {
     return new RustFile(path, TokenTree.of(content));
   }
 
-  private RustFile unionConversionModule(final UnionShape unionShape) {
-    final Map<String, String> variables = MapUtils.merge(
-      serviceVariables(),
-      unionVariables(unionShape)
-    );
-
-    final List<Map<String, String>> perMemberVariables = unionShape
-      .members()
-      .stream()
-      .map(memberShape -> {
-        final Map<String, String> memberVariables = MapUtils.merge(
-          variables,
-          unionMemberVariables(memberShape)
-        );
-        final Shape targetShape = model.expectShape(memberShape.getTarget());
-        memberVariables.put(
-          "innerToDafny",
-          toDafny(targetShape, "x", false, false).toString()
-        );
-        memberVariables.put(
-          "innerFromDafny",
-          fromDafny(targetShape, "x", false, false).toString()
-        );
-        return memberVariables;
-      })
-      .toList();
-
-    variables.put(
-      "toDafnyVariants",
-      perMemberVariables
-        .stream()
-        .map(memberVariables ->
-          IOUtils.evalTemplate(
-            """
-            $qualifiedRustUnionName:L::$rustUnionMemberName:L(x) =>
-                crate::r#$dafnyTypesModuleName:L::$dafnyUnionName:L::$dafnyUnionMemberName:L {
-                    $dafnyUnionMemberName:L: $innerToDafny:L,
-                },
-            """,
-            memberVariables
-          )
-        )
-        .collect(Collectors.joining("\n"))
-    );
-    variables.put(
-      "fromDafnyVariants",
-      perMemberVariables
-        .stream()
-        .map(memberVariables ->
-          IOUtils.evalTemplate(
-            """
-            crate::r#$dafnyTypesModuleName:L::$dafnyUnionName:L::$dafnyUnionMemberName:L {
-                $dafnyUnionMemberName:L: x @ _,
-            } => $qualifiedRustUnionName:L::$rustUnionMemberName:L($innerFromDafny:L),
-            """,
-            memberVariables
-          )
-        )
-        .collect(Collectors.joining("\n"))
-    );
-
-    final String content = IOUtils.evalTemplate(
-      getClass(),
-      "runtimes/rust/conversions/union.rs",
-      variables
-    );
-    final Path path = Path.of(
-      "src",
-      "conversions",
-      "%s.rs".formatted(toSnakeCase(unionName(unionShape)))
-    );
-    return new RustFile(path, TokenTree.of(content));
-  }
-
   private RustFile resourceConversionModule(final ResourceShape resourceShape) {
     final Map<String, String> variables = MapUtils.merge(
       serviceVariables(),
@@ -1296,142 +1222,6 @@ public class RustLibraryShimGenerator extends AbstractRustShimGenerator {
       rustTypeForShape(model.expectShape(memberShape.getTarget()))
     );
     return variables;
-  }
-
-  protected String unionName(final UnionShape unionShape) {
-    return unionShape.getId().getName(service);
-  }
-
-  protected String rustUnionName(final UnionShape unionShape) {
-    return toPascalCase(unionName(unionShape));
-  }
-
-  protected String qualifiedRustUnionName(final UnionShape unionShape) {
-    return "%s::%s".formatted(
-        getRustTypesModuleName(),
-        rustUnionName(unionShape)
-      );
-  }
-
-  /**
-   * Generates values for variables commonly used in union-specific templates.
-   */
-  protected HashMap<String, String> unionVariables(
-    final UnionShape unionShape
-  ) {
-    final HashMap<String, String> variables = new HashMap<>();
-    final String unionName = unionName(unionShape);
-    variables.put("unionName", unionName);
-    variables.put("snakeCaseUnionName", toSnakeCase(unionName));
-    variables.put("dafnyUnionName", unionName);
-    variables.put("rustUnionName", rustUnionName(unionShape));
-    variables.put("qualifiedRustUnionName", qualifiedRustUnionName(unionShape));
-    return variables;
-  }
-
-  protected String unionMemberName(final MemberShape memberShape) {
-    return memberShape.getMemberName();
-  }
-
-  protected String rustUnionMemberName(final MemberShape memberShape) {
-    return toPascalCase(unionMemberName(memberShape));
-  }
-
-  protected String dafnyUnionMemberName(final MemberShape memberShape) {
-    return unionMemberName(memberShape);
-  }
-
-  /**
-   * Generates values for variables commonly used in union-member-specific templates.
-   */
-  protected HashMap<String, String> unionMemberVariables(
-    final MemberShape memberShape
-  ) {
-    final String memberName = unionMemberName(memberShape);
-    final Shape targetShape = model.expectShape(memberShape.getTarget());
-
-    final HashMap<String, String> variables = new HashMap<>();
-    variables.put("unionMemberName", memberName);
-    variables.put("snakeCaseUnionMemberName", toSnakeCase(memberName));
-    variables.put("dafnyUnionMemberName", dafnyUnionMemberName(memberShape));
-    variables.put("rustUnionMemberName", rustUnionMemberName(memberShape));
-    variables.put("unionMemberType", rustTypeForShape(targetShape));
-    return variables;
-  }
-
-  private String rustTypeForShape(final Shape shape) {
-    return switch (shape.getType()) {
-      case BOOLEAN -> "::std::primitive::bool";
-      // integral
-      case BYTE -> "::std::primitive::i8";
-      case SHORT -> "::std::primitive::i16";
-      case INTEGER -> "::std::primitive::i32";
-      case LONG -> "::std::primitive::i64";
-      // floats
-      case FLOAT -> "::std::primitive::f32";
-      case DOUBLE -> "::std::primitive::f64";
-      // special numerics
-      case BIG_INTEGER -> "::num::bigint::BigInt";
-      case BIG_DECIMAL -> "::num::rational::BigRational";
-      // special collections
-      case BLOB -> "::aws_smithy_types::Blob";
-      case STRING -> {
-        //noinspection deprecation
-        if (shape.hasTrait(EnumTrait.class)) {
-          yield qualifiedRustEnumType(
-            ModelUtils.stringToEnumShape(shape.asStringShape().orElseThrow())
-          );
-        }
-        yield "::std::string::String";
-      }
-      case ENUM -> qualifiedRustEnumType(shape.asEnumShape().orElseThrow());
-      // other simple shapes
-      case TIMESTAMP -> "::aws_smithy_types::DateTime";
-      // aggregates
-      case STRUCTURE -> {
-        final StructureShape structureShape = (StructureShape) shape;
-        if (shouldGenerateStructForStructure(structureShape)) {
-          yield qualifiedRustStructureType(structureShape);
-        }
-        if (structureShape.hasTrait(ReferenceTrait.class)) {
-          yield qualifiedRustReferenceType(structureShape);
-        }
-        throw new UnsupportedOperationException(
-          "Unsupported type for structure: " + shape.getId()
-        );
-      }
-      case LIST -> {
-        final ListShape listShape = (ListShape) shape;
-        final String memberType = rustTypeForShape(
-          model.expectShape(listShape.getMember().getTarget())
-        );
-        yield "::std::vec::Vec<%s>".formatted(memberType);
-      }
-      case MAP -> {
-        final MapShape mapShape = (MapShape) shape;
-        final String keyType = rustTypeForShape(
-          model.expectShape(mapShape.getKey().getTarget())
-        );
-        final String valueType = rustTypeForShape(
-          model.expectShape(mapShape.getValue().getTarget())
-        );
-        yield "::std::collections::HashMap<%s, %s>".formatted(
-            keyType,
-            valueType
-          );
-      }
-      case UNION -> {
-        if (ModelUtils.isInServiceNamespace(shape.getId(), service)) {
-          yield qualifiedRustUnionName((UnionShape) shape);
-        }
-        throw new UnsupportedOperationException(
-          "Unsupported type for union: " + shape.getId()
-        );
-      }
-      default -> throw new UnsupportedOperationException(
-        "Unsupported shape type: " + shape.getType()
-      );
-    };
   }
 
   protected boolean isRustFieldRequired(
